[db/backends] Extract InfluxQL base client
Move the InfluxQL DatabaseClient implementation into a shared base class and make the InfluxDB backend client a thin subclass.

This refactor paves the way for adding an InfluxDB 2.x backend in InfluxQL compatibility mode without duplicating the client implementation.

No behavior changes expected; tests and ./run-qa-checks pass.
